finally, made text layout closely matches browser

+++ radiant/layout_table.cpp
@@ -3,7 +3,7 @@
 #include "../lib/log.h"
 
 // Forward declaration for CSS length resolution
-int resolve_length_value(LayoutContext* lycon, uintptr_t property,
+float resolve_length_value(LayoutContext* lycon, uintptr_t property,
     const lxb_css_value_length_percentage_t *value);
 
 /*
